add expenses crud for stockadd correction.

+++ src/main/java/com/stuffhouse/myapp/domain/Expenses.java
@@ -20,7 +20,7 @@ public class Expenses {
     private String type;
     private long quantity;
     private String description;
-    private double cost;
+    private long cost;
 
 
 }
